manager refactor done, waiting for test
