+++ include/Drivers/MotorController.hpp
@@ -10,23 +10,24 @@ enum MotorMode { TORQUE, SPEED, VELOCITY };
 
 class MotorController {
 public:
-    MotorController() {};
+  MotorController(){};
   virtual void begin(long speed);
-//   virtual void setMode(uint8_t channel, MotorMode mode) = 0;
+  //   virtual void setMode(uint8_t channel, MotorMode mode) = 0;
   virtual int32_t getSpeed(uint8_t address, uint8_t channel, bool &valid) = 0;
   virtual void setSpeed(int32_t speed, uint8_t address, uint8_t channel, bool &valid) = 0;
-//   virtual int32_t getPosition(uint8_t channel) = 0;
-//   virtual void setPosition(uint8_t channel) = 0;
+  //   virtual int32_t getPosition(uint8_t channel) = 0;
+  //   virtual void setPosition(uint8_t channel) = 0;
 };
 
 class RoboClawController : public MotorController {
 public:
-    RoboClawController(RoboClaw &roboclaw) : _roboclaw(roboclaw) {};
-    void begin(long speed);
-    int32_t getSpeed(uint8_t address, uint8_t channel, bool &valid) override;
-    void setSpeed(int32_t speed, uint8_t address, uint8_t channel, bool &valid) override;
+  RoboClawController(RoboClaw &roboclaw) : _roboclaw(roboclaw){};
+  void begin(long speed);
+  int32_t getSpeed(uint8_t address, uint8_t channel, bool &valid) override;
+  void setSpeed(int32_t speed, uint8_t address, uint8_t channel, bool &valid) override;
+
 private:
-    RoboClaw &_roboclaw;
+  RoboClaw &_roboclaw;
 };
 
 }  // namespace motors
