+++ src/Drivers/EthernetDriver.cpp
@@ -1,15 +1,15 @@
-#include "EthernetDriver.hpp"
+#include "Drivers/EthernetDriver.hpp"
 
 namespace ethernet {
 
 elapsedMillis EthernetDriver::_sendTimer;
 
 bool EthernetDriver::sendTimeHasElapsed() {
-    return _sendTimer >= TIMER_DURATION_MS;
+  return _sendTimer >= TIMER_DURATION_MS;
 }
 
 void EthernetDriver::resetSendTimer() {
-    _sendTimer -= TIMER_DURATION_MS;
+  _sendTimer -= TIMER_DURATION_MS;
 }
 
-}
+}  // namespace ethernet
